Add hacked formula for mosh w/ local scrollback
- From here:
  https://github.com/mobile-shell/mosh/compare/master...rledisez:localScrollback-1.3.2?diff=unified
  and
  https://donnlee.com/2018/03/31/mosh-with-iterm2s-tmux-integration/

First cut, hacked mosh works but fails tests.

This adds the patch to enable local scrollback, I'm using it and it
seems to work.

The build fails its tests though which isn't surprising, since all of
the terminal emulation bits stuff are no longer hooked up.

I'll clean it up after I use it for a while and see how it works.